requirements.txt
discord.py=1.2.4
requests=2.24.0